fix: remove abort controller deps
`AbortController` and `AbortSignal` are in browsers and LTS node so the `abort-controller` and `native-abort-controller` deps aren't required any more.

I had to access `AbortController` as a property of `globalThis` as otherwise `standard` gets upset and there's no easy way to configure it without adding a whole load of extra deps to this project.